+++ CHANGELOG.md
@@ -17,9 +17,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Harmonize how cards look on the site (grey border and white background),
+- Move all the code demo site related to the demo site to its own application,
 - Activating a filter that is a parent or child of a current active filter
   removes this active relative. This makes the experience of adding those
-  relative filters more intuitive.
+  relative filters more intuitive,
 - Simplify Richie settings and provide defaults for those unlikely to be
   customized (search, languages, plugins, general).
