+++ src/richie/apps/demo/fixtures/logo/logo1.png
