+++ src/richie/apps/demo/fixtures/logo/logo2.png
