+++ src/richie/apps/demo/fixtures/logo/logo3.png
