+++ src/richie/apps/demo/fixtures/logo/logo4.png
