+++ src/richie/apps/demo/static/images/logo.png
